Remove master outlet, doesn't work in IP.
The master outlet is not properly defined in both HTTP and IP, I should have verified that in the IP driver, so this resolves that quirk. Also re-ran the black formatter and noticed that there was a missing unit so I added that. I removed the suggested unit because there is no viable converter for duration.